requirements.txt
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu== 1.7.4
altair<5
unstructured==0.7.12
unstructured-client
python-dotenv==1.0.0
langchain==0.0.154
langchain-community==0.0.37
langchain-core==0.1.52
langchain-text-splitters==0.0.1
openai==0.27.8
streamlit-extras